api/requirements.txt
appdirs==1.4.4
astroid==2.5.1
autopep8==1.5.6
certifi==2020.12.5
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
pipenv==2020.11.15
protobuf==3.15.8
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pylint==2.7.2
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.4.11
toml==0.10.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1